pyproject.toml
[project]
name = "vlm-comparison-tool"
version = "0.1.0"
description = "VLM Comparison Tool with Multimodal Retrieval"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.38.2",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "requests>=2.32.4",
    "byaldi>=0.0.5",
    "psutil>=6.1.0",
    "python-dotenv>=1.0.0",
    "torch>=2.6.0,<2.7.0",
    "transformers>=4.51.1,<4.52.0",
    "gputil>=1.4.0",
    "skypilot[gcp]>=0.10.1.post1",
]
